Rename hasp_ethernet driver for stm32
