+++ src/hasp_ethernet_stm32.h
@@ -1,8 +1,8 @@
 /* MIT License - Copyright (c) 2020 Francis Van Roie
    For full license information read the LICENSE file in the project folder */
 
-#ifndef HASP_ETHERNET_H
-#define HASP_ETHERNET_H
+#ifndef HASP_ETHERNET_STM32_H
+#define HASP_ETHERNET_STM32_H
 
 void ethernetSetup();
 void ethernetLoop(void);
